Renamed ConnectWebSocket as useSocket
This was done to indicate that it is a hook (following React's
strong recommendation that hooks be named with the "use" prefix) and to
remove references to WebSockets, as the implementation might be changed
to use the SocketIO library in the future.

Also, the hook now returns the the current node of the useRef hook, so
that it is not necessary to call current in other files using the
useSocket hook.